update: move mcm to gdinn file
